Add dependency cycle checks to get_target_dependencies() cmake function
This commit adds a check to the get_target_dependencies() cmake function
to check if the target has a dependency cycle, avoiding an infinite
recursion in the function. When detected, the function aborts CMake
execution with a fatal error, and the full dependency chain is shown.

The full dependency walking can be debugged by setting the CMake
variable TINYREFL_UTILS_DEBUG_GETTARGETDEPENDENCIES variable to ON. This
will add traces to the CMake config output while walking the dependency
tree.