Conditionally run static tests depending on array subscript
For some reason GCC 5 does not evaluate array and pointer subscripts as
constant expressions, so features such as the set of arguments of an
attribute cannot be evaluated at compile time with that compiler. I have
added some cmake checks and macros to ensure the library is aware of
this corner case.