Make sure all GCC images export the right gcc/g++ version as CC/CXX
Some of these images have gcc/g++ not aliased to the right
version installed in the images (i.e. the image gcc8 has gcc-8 installed
but gcc is gcc-7, not an alias to gcc-8). Since conan uses CC
environment variable to set the profile settings.compiler, we need
to make sure CC is aliased to the right compiler version