Fix to_json()/from_json() dispatching
JSON For Modern C++ implents integration of User Defined Types (UDTs)
through ADL. The nlohmann::detail::has_to_json family of traits does not
check that but that the serializer picked by the JSON type
(adl_serializer by default) is compatible with the UDT. This check does
not actually check if there's an ADL overload available, so that always
returns true.

> The check is actually done later through SFINAE when invoking the
from/to_json() functions.

Now we correctly check all the combinations of:

 - We have reflection metadata for the UDT
 - There is an ADL overload of from/to_json() declared for the UDT

and dispatch the calls to our reflection-based serializers or
JSON4ModernC++ lib accordingly.